Create font family per PPPangrams to support different font weights
Currently the PPPangrams font files all claim to have the same fontWeight,
there doesn't seem to be a way to override this in flutter so we have to
have different font families per font. In an ideal world the fonts would
be updated and then we could go back to using font weight directly.